Move ImageViewManager to its own folder
