+++ vnext/ReactUWP/Base/UwpReactInstance.cpp
@@ -34,7 +34,7 @@
 #include <Views/TextViewManager.h>
 #include <Views/VirtualTextViewManager.h>
 #include <Views/ViewViewManager.h>
-#include <Views/ImageViewManager.h>
+#include <Views/Image/ImageViewManager.h>
 #include <Views/WebViewManager.h>
 
 // Polyester View Managers // TODO: Move Polyester implementations out of this library and depot
