+++ vnext/ReactUWP/ReactUWP.vcxproj.filters
@@ -91,9 +91,6 @@
     <ClCompile Include="Views\FrameworkElementViewManager.cpp">
       <Filter>Views</Filter>
     </ClCompile>
-    <ClCompile Include="Views\ImageViewManager.cpp">
-      <Filter>Views</Filter>
-    </ClCompile>
     <ClCompile Include="Views\RawTextViewManager.cpp">
       <Filter>Views</Filter>
     </ClCompile>
@@ -216,6 +213,9 @@
     <ClCompile Include="Views\ScrollContentViewManager.cpp">
       <Filter>Views</Filter>
     </ClCompile>
+    <ClCompile Include="Views\Image\ImageViewManager.cpp">
+      <Filter>Views\Image</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Modules\DeviceInfoModule.h">
@@ -304,9 +304,6 @@
     </ClInclude>
     <ClInclude Include="ABI\HStringHelper.h">
       <Filter>ABI</Filter>
-    </ClInclude>
-    <ClInclude Include="Views\ImageViewManager.h">
-      <Filter>Views</Filter>
     </ClInclude>
     <ClInclude Include="Views\RawTextViewManager.h">
       <Filter>Views</Filter>
@@ -431,6 +428,9 @@
     <ClInclude Include="Views\Impl\SnapPointManagingContentControl.h" />
     <ClInclude Include="Views\ScrollContentViewManager.h">
       <Filter>Views</Filter>
+    </ClInclude>
+    <ClInclude Include="Views\Image\ImageViewManager.h">
+      <Filter>Views\Image</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -559,6 +559,9 @@
     </Filter>
     <Filter Include="Views\Impl">
       <UniqueIdentifier>{101c2d4a-d343-4771-969f-29af0c9f46b2}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Views\Image">
+      <UniqueIdentifier>{6d877886-1c5c-45b8-8eae-2e1354cf673e}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
