+++ vnext/ReactUWP/Views/Image/ImageViewManager.cpp
