+++ vnext/ReactUWP/Views/Image/ImageViewManager.h
